Remove circular dependencies from commons
