+++ packages/commons/src/pdf-generator/pdfGenerator.ts
@@ -3,7 +3,7 @@ import path from "path";
 import { fileURLToPath } from "url";
 import { pdfGenerationError } from "pagopa-interop-models";
 import puppeteer, { Browser } from "puppeteer";
-import { buildHTMLTemplateService } from "../index.js";
+import { buildHTMLTemplateService } from "../templating/htmlTemplateService.js";
 
 export interface PDFGenerator {
   generate: (
@@ -15,8 +15,8 @@ export interface PDFGenerator {
 export async function initPDFGenerator(): Promise<PDFGenerator> {
   const templateService = buildHTMLTemplateService();
   let browserInstance = await puppeteer.launch({
-    /* NOTE 
-      those configurations allow link (file://) usages for 
+    /* NOTE
+      those configurations allow link (file://) usages for
       resources files in template's folder
     */
     args: [
